Refactor: 랭킹 페이지 파일 구조 수정 및 불필요한 코드 정리
