--- src/pages/Ranking/RankContents.module.css
@@ -1,15 +1,0 @@
-.container {
-  padding: 0 var(--container-m-sm, 12px);
-  margin-block-end: 60px;
-}
-
-.rankCardList {
-  display: flex;
-  flex-direction: column;
-  gap: var(--space-10, 40px);
-  margin-block-end: var(--space-10, 40px);
-}
-
-.error {
-  color: var(--text-300, #6e6e6e);
-}
